requirements.txt
matplotlib==3.8.0
mplcursor==0.6.1